+++ lib/tflite_audio.dart
@@ -6,17 +6,17 @@ import 'package:flutter/services.dart';
 class TfliteAudio {
   static const MethodChannel _channel = MethodChannel('tflite_audio');
 
-  static Future<dynamic> startAudioRecognition() async {
-    return _channel.invokeMethod('startAudioRecognition');
+  static Future<dynamic> startAudioRecognition(
+      int sampleRate, int recordingLength, int bufferSize) async {
+    return _channel.invokeMethod('startAudioRecognition', {
+      'sampleRate': sampleRate,
+      'recordingLength': recordingLength,
+      'bufferSize': bufferSize
+    });
   }
 
-  // ignore: type_annotate_public_apis
   static Future loadModel(
-      // ignore: avoid_positional_boolean_parameters
-      String model,
-      String label,
-      int numThreads,
-      bool isAsset) async {
+      String model, String label, int numThreads, bool isAsset) async {
     return _channel.invokeMethod(
       'loadModel',
       {