+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 
+## 0.0.4
+* Added the following arguments into the future: startAudioRecognition()
+    - sampleRate
+    - recordingLength
+    - bufferSize
+
 ## 0.0.3
-Merged permission and audio recognition futures into one future.
+* Merged permission and audio recognition futures into one future.
 
 ## 0.0.2
-Fixed image url
+* Fixed image url
 
 ## 0.0.1